+++ src/FreeElephants/RestDaemon/Endpoint/Handler/AbstractEndpointMethodHandler.php
@@ -3,7 +3,6 @@
 namespace FreeElephants\RestDaemon\Endpoint\Handler;
 
 use FreeElephants\RestDaemon\Endpoint\EndpointInterface;
-use FreeElephants\RestDaemon\Endpoint\EndpointMethodHandlerInterface;
 use FreeElephants\RestDaemon\Middleware\Collection\EndpointMiddlewareCollectionInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;