+++ src/FreeElephants/RestDaemon/Endpoint/Handler/DefaultHandlerFactory.php
@@ -4,8 +4,6 @@
 namespace FreeElephants\RestDaemon\Endpoint\Handler;
 
 
-use FreeElephants\RestDaemon\Endpoint\EndpointMethodHandlerInterface;
-
 class DefaultHandlerFactory implements HandlerFactoryInterface
 {