+++ src/FreeElephants/RestDaemon/Endpoint/Handler/InjectionHandlerFactory.php
@@ -5,7 +5,6 @@ namespace FreeElephants\RestDaemon\Endpoint\Handler;
 
 
 use FreeElephants\DI\Injector;
-use FreeElephants\RestDaemon\Endpoint\EndpointMethodHandlerInterface;
 
 class InjectionHandlerFactory implements HandlerFactoryInterface
 {